### 2024-02-25 #238 @editorialbot
[PRE REVIEW]: Static Equilibrium in Remote Alaska
**Submitting author:** @rkn2 (<a href="http://orcid.org/0000-0002-8939-5998">Rebecca Napolitano</a>)
**Repository:** https://github.com/rkn2/AK_statics
**Branch with paper.md** (empty if default branch):
**Version:** v1
