### 2025-04-16 #238 @nicoguaro
@ntmoore, @satish-annigeri are you willing to review this work about teaching statics?
