### 2025-08-07 #301 @nicoguaro
@rkn2, could you check the missing DOIs in the references and the captions for your figure and table?
