### 2025-08-07 #301 @nicoguaro
@dsaninv2, @Tejada90, this is the space where the review process takes form. There is a checklist for each one; tick the boxes when you see that the criterion is satisfied. You can generate the checklist with

I will be here to answer the questions that you might have.

Let us set the last week of August as a tentative deadline. Is that OK for you?
