### 2025-09-02 rkn2/AK_statics#1 @dsaninv2
JOSE Review First Round
# Statics_Lesson1.ipynb – Review Comments

## Strengths
- Excellent alignment with Common Core standards; makes the lesson credible and integrable in formal curricula.
- Strong contextualization: links physics/engineering concepts with Alaskan cultural, social, and environmental realities.
- Scaffolded learning: progresses from hand calculation → Python coding → visualization with plots.
- Vocabulary section is clear and helpful for first-year high school students.

## Issues / Weaknesses
- Minor typos: “Sutdents” → Students, “analyize” → analyze.
- Video placeholder: The code block for `IPython.display.HTML` does not contain a working link (currently just `video_html = """ """`). Replace with actual embedded video or remove.
- Units: Constants use PSI and inches, while students should also be familiar with SI (MPa, meters). At least provide dual-unit examples.
- Code readability: Some explanations are verbose. Students may benefit from concise comments plus a short “summary cell” that shows the complete workflow at once.

## Suggestions for Improvement
- Learning Objectives: Reframe as measurable outcomes (e.g., “Students will be able to calculate safe ice loads using Python functions”).
- Code clarity:
  - Show a single-cell compact version after the step-by-step walkthrough, so students see the “full solution” cleanly.
  - Use f-strings consistently for output (`print(f"Ice Thickness: {h} in, Safe Load: {safe_load} lbs")`).
- Plot improvements: Add grid lines, units on axes, and compare safe vs. theoretical load.
- Assessment prompts: Add quick tasks at the end (“What happens if thickness = 8 inches?” or “Modify safety factor to 3, what changes?”).
- Context connection: After calculations, explicitly ask students to reflect: “How might climate change affect the assumptions in our model?”

---

# Statics_Lesson2 (1).ipynb – Review Comments

## Strengths
- Clear learning objectives that separate equilibrium of forces from torque (appropriate scaffolding).
- Good use of step-by-step reasoning: equilibrium equations → coding → plotting.
- Visualization: Plot comparing support forces vs. ice capacity is effective for conceptual learning.
- Nicely emphasizes sign conventions, which many students struggle with.

## Issues / Weaknesses
### Notation & Math Presentation
- Symbols for forces are inconsistent: sometimes FN1/FN2, sometimes S1/S2. Stick to one.

### Code
- The stability check `if FN1 <= P_capacity`: only tests one support force, not the total. A case where each foot load < capacity but combined > capacity is not flagged. This should be clarified or modified.
- Redundant explanations in comments (how loops work) make code cells long; suggest a concise “annotated” version plus a clean minimal version for clarity.
- Use f-strings consistently for readable output instead of mixed formats.

### Pedagogical Issues
- Example is simple (equal load distribution). For more substantial conceptual value, show a non-symmetric case where forces differ.
- The concept check section is good but could include a direct student coding task (“Modify the weights list to include 600 lb and discuss stability”).
- The note on solving for unknowns is text-heavy. Could be condensed with one symbolic worked example (system of equations with substitution).

## Suggestions for Improvement
- Standardize variable naming (use FN1, FN2 throughout).
- Revise stability check: compare total weight vs. capacity, not only one support force.
- Add at least one non-symmetric example (person carrying a load more on one foot).
- Provide a clean solution cell at the end of each example to reinforce the workflow.

---

# Statics_Lesson3.ipynb – Review Comments

## Strengths
- Clear Learning Objectives:
  - The objectives articulate both conceptual understanding (equilibrium, torque, lever arm) and application (real-world torque calculations).
  - Explicit connection between translational and rotational equilibrium is strong.
- Identification of Sticking Points:
  - Preemptively flagging sign conventions, torque vs moment, and moment connections is excellent. Anticipating misconceptions helps students avoid confusion.
- Vocabulary Section:
  - Key terms are clearly defined, with contextual examples.
  - The distinction between stable and unstable equilibrium is well explained with intuitive physical analogies.
- Use of Real-World Examples:
  - Crane and fish example is relatable, makes the abstract idea of torque tangible.
  - Safety considerations (manufacturer torque vs theoretical) introduce professional engineering practice.
- Coding Integration:
  - Python snippets reinforce equations numerically, providing a computational angle.
  - The inclusion of "teacher’s choice" options (outline vs full code) promotes adaptable teaching strategies.
- Progression of Examples:
  - Example 1 introduces torque in a practical crane problem.
  - Example 2 ties back to the previous lesson, showing the evolution of knowledge.
  - The stability/balance section integrates physics intuition with statics principles.

## Areas for Improvement
- Clarity and Grammar:
  - Multiple typos (e.g., calcualte, undersand, equilbrium, intereseted, direciton, raotate). These may distract students.
- Mathematical Rigor:
  - In Example 1, the explanation of why the torque is not zero could be more formal: explicitly show that the moment connection introduces an unknown reaction torque, leading to equilibrium equation.
- Coding Pedagogy:
  - Current Python code is correct but lacks comments linking to equations. Students might struggle to connect code with theory.
  - Suggestion: add intermediate print statements (e.g., lever arm in ft, fish torque in lbft).
- Safety and Design Context:
  - The note about the crane’s actual load rating (220 lb vs 807 lb theoretical) is excellent, but it could be tied more explicitly to factor of safety concepts.

## Recommendations
- Enhance Coding Cells:
  - Provide step-by-step code outlines more explicitly tied to equations.
  - Example:
    ```python
    # Step 1: Define known quantities
    FW = 1000  # weight of fish [lb]
    arm_distance = 46/12  # lever arm [ft]

    # Step 2: Calculate torque
    tau_fish = FW * arm_distance

    # Step 3: State result
    print("Torque due to fish load at point A:", tau_fish, "lb·ft")
    ```
- Link Back to Engineering Practice:
  - When introducing manufacturer limits, explicitly define factor of safety (FoS) and ask students to compute it.

---

# Statics_Lesson4.ipynb – Review Comments

## Strengths
- Clear step-by-step breakdown of how to approach a statics problem, which helps students structure their thinking.
- Good use of real-world examples (truck and trailer on ice) that make the abstract concepts more relatable.
- Consistent inclusion of visuals, coding activities, and flowcharts to appeal to different learning styles.

## Issues / Weaknesses
- The vocabulary section says “no new vocabulary” but the lesson still has terms like pivot point and support force that could use a short definition.
- The pivot point explanation is correct but too abstract; many students may not see why A is chosen beyond "to simplify."
- Code is too complete. Students mostly just "run" instead of reasoning. Little space for discovery.
- Heavy reliance on hard-coded numbers (1500, 2, 5, etc.), no user input or variability.
- No error checks (e.g., negative weights, unrealistic values).
- Plotting is done but lacks labels, gridlines, or legend clarity for real interpretability.
- Incremental coding tasks ("delete or keep") are confusing; students may not know which version to use.

## Suggestions for Improvement
- Add short definitions for recurring terms (pivot, support force, torque) even if they were introduced earlier.
- Provide optional coding scaffolds: one full example, one partial outline, and one blank cell for different skill levels.
- Emphasize why each step matters (choosing the pivot point simplifies solving, not just “because it’s natural”).
- In the results section, tie the numbers back to real-world meaning (what failure would look like on the ice).
- Encourage experimentation: e.g., "Change trailer weight or ice thickness and see what happens."

---

# Statics_Lesson5.ipynb – Review Comments

## Strengths
- Real-world context: Using bridges makes the lesson engaging and relevant for students.
- Progressive complexity: Moves from simple two-abutment analysis to internal forces and then to the effect of adding piers. This scaffolding is effective.
- Integration of coding: Students are encouraged to model different load conditions computationally, which reinforces learning.
- Class activities and reflections: Promotes critical thinking about engineering tradeoffs (safety, cost, environment).
- Instructor notes: Adding ecological, navigational, and cost considerations broadens the scope beyond pure mechanics.

## Issues / Weaknesses
### Pedagogical
1. Abstract jump: The concept of “infinitely close cut” for internal forces is introduced abruptly with little intuitive explanation. Students may struggle to understand why we “cut” the structure and what that reveals.
2. Heavy reliance on equations: Long symbolic sections dominate before intuition is reinforced. Visuals help, but students may still lose track.
3. Uneven treatment of vocabulary: Terms like shear, moment, and cut section are assumed to be known rather than reinforced with simple analogies.
4. Overloaded lesson: Covers abutments, piers, shear, moment, cuts, and environmental tradeoffs in a single notebook. Risk of cognitive overload.

### Code
1. Too complete: Most code cells already provide full solutions, leaving little room for students to practice problem-solving.
2. Hard-coded values: Constants like 600,000 lb and 300 ft are fixed; this discourages exploration of alternative scenarios.

## Suggestions for Improvement
### Pedagogy
- Introduce “cut sections” with a physical analogy (like sawing a plank and holding one piece) to make the rationale intuitive.
- Split lesson into two notebooks: one on external equilibrium (abutments, piers) and one on internal forces (shear and moment).
- Add a brief recap glossary for shear, moment, cut section, and equilibrium to avoid assumed knowledge.
- Use more conceptual questions before equations: for instance: “If a truck parks closer to one abutment, which support feels more load?”

### Code
- Provide scaffolded coding tasks: start with outlines, then let students complete missing steps instead of giving full solutions.
- Replace repeated hard-coded problems with functions (e.g., `def support_forces(length, load, position): ...`).
- Add interactive sliders (via ipywidgets) for bridge length, weight, and load position to make the exploration more engaging and fun.
- Make expectations clear: consistently present each coding task as (a) full worked example, (b) partial outline, or (c) blank scaffold.
