### 2025-09-02 #301 @dsaninv2
# General Review of AK_statics Repository

**Note:** Specific detailed comments on individual lessons are available in [Issue #1](https://github.com/rkn2/AK_statics/issues/1).

---

## General Evaluation
This manuscript presents an innovative, context-rich physics module that integrates fundamental concepts of statics with computational practice in Python, situated within the social and environmental realities of rural Alaska. The work is timely, creative, and strongly motivated by both pedagogical and societal needs. It provides not only a technical introduction to forces, equilibrium, and torque for high school students, but also situates physics within the lived experiences of communities facing climate and infrastructure challenges. This makes the project a compelling example of interdisciplinary STEM education with social relevance.

The repository is clearly organized, with lessons structured as Google CoLab notebooks, supplemented by teacher-oriented resources. The writing is accessible yet rigorous, and the documentation highlights both the physics content and its broader implications. The inclusion of teacher and student feedback is a valuable addition that demonstrates the module’s practical effectiveness.

I find this to be a strong contribution, and I encourage publication after considering some refinements that could strengthen clarity, pedagogy, and technical consistency.

---

## Strengths
- Embedding statics within the realities of Alaskan communities, including issues of ice safety and bridge infrastructure, gives students a meaningful and motivating frame for learning.
- The blending of physics, engineering, programming, and social sciences enriches the material and encourages students to see STEM as deeply connected with society.
- Lessons gradually build from basic forces to torque and bridge equilibrium, scaffolding student learning.
- Using Python throughout provides an excellent introduction to computational methods in physics.
- Hosting on CoLab enables both classroom and self-study use with minimal setup.
- The inclusion of code outlines and usage guidance helps teachers adopt and adapt the materials.

---

## Areas for Improvement

### Pedagogy
1. **Balance of Scope and Depth**
   Several lessons (particularly Lessons 4 and 5) introduce many new concepts in a single notebook (torque, pivot choice, internal forces, cut sections). This can risk cognitive overload. Consider splitting some content or emphasizing conceptual overviews before detailed calculations.

2. **Sticking Points**
   Concepts such as torque sign convention and internal forces via “cut sections” are potential stumbling blocks. More intuitive analogies or physical demonstrations (e.g., cutting and holding a beam) would help ground these abstractions.

3. **Vocabulary Reinforcement**
   Some important terms (shear force, internal moment) are introduced but not consistently revisited. A running glossary or periodic “concept checks” could reinforce understanding.

4. **Student Differentiation**
   The coding content is sometimes too detailed for experienced learners but too complete for novices (with many cells providing full solutions). Offering two tracks, one with full code and one with partial scaffolds, would address the diversity of student backgrounds.

### Code and Technical Materials
1. **Over-Completion of Code**
   Many examples already contain full solutions, leaving little room for exploration. Stronger use of scaffolded pseudocode or fill-in-the-blank style tasks could enhance student engagement.

2. **Repetition Without Abstraction**
   Several calculations repeat the same structure (support reactions at varying distances) without generalizing to a function or interactive tool. Encapsulating these in functions (e.g., `support_forces(length, load, position)`) would both model good coding practices and reduce redundancy.

3. **Visualization Opportunities**
   While equilibrium of forces is visualized, internal forces such as shear and moment could be shown graphically in videos or animations (shear/moment diagrams), which are central to structural analysis.

4. **Interactivity**
   The use of `ipywidgets` sliders for load position, ice thickness, or bridge length would allow students to experiment more dynamically, strengthening conceptual understanding.

### Manuscript and Repository Presentation
1. **Minor Formatting**
   Ensure equations, code, and narrative are consistently styled. Occasional formatting irregularities (symbols, spacing, or image placeholders) could be polished for readability.

---

## Suggestions for Enhancement
- Divide Lesson 5 (bridges) into two parts: external equilibrium and internal forces, so that students can fully process each.
- Provide an optional condensed coding track for learners with prior programming knowledge.
- Add interactive elements in Python (sliders for load position, thickness, etc.) to make the notebooks more exploratory.
- Include a short assessment appendix with example quiz questions or exercises that could be used by teachers to measure learning outcomes.

---

## Recommendation
This project is a creative and socially meaningful contribution to physics education. It succeeds in combining fundamental statics, computational skills, and awareness of real-world challenges in Alaska. With some refinements to lesson scope, coding scaffolding, and visual interactivity, the module could serve as a model for contextualized STEM education in other regions as well.

**I recommend publication after minor revisions** and commend the authors for developing a resource that will not only teach physics but also inspire students to connect science with community needs and environmental responsibility.

> ## Review checklist for [@dsaninv2](https://github.com/dsaninv2)
> ### Conflict of interest
> * [x]  As the reviewer I confirm that I have read the [JOSE conflict of interest policy](https://github.com/openjournals/joss/blob/jose/COI.md) and that there are no conflicts of interest for me to review this work.
>
> ### Code of Conduct
> * [x]  I confirm that I read and will adhere to the [JOSE code of conduct](https://jose.theoj.org/about#code_of_conduct).
>
> ### General checks
> * [x]  **Repository:** Is the source for this learning module available at the https://github.com/rkn2/AK_statics?[x]  **License:** Does the repository contain a plain-text LICENSE file with the contents of a standard license? (OSI-approved for code, Creative Commons for content)[x]  **Version:** Does the release version given match the repository release?[x]  **Authorship:** Has the submitting author ([@rkn2](https://github.com/rkn2)) made visible contributions to the module? Does the full list of authors seem appropriate and complete?
>
> ### Documentation
> * [x]  **A statement of need:** Do the authors clearly state the need for this module and who the target audience is?[x]  **Installation instructions:** Is there a clearly stated list of dependencies?[x]  **Usage:** Does the documentation explain how someone would adopt the module, and include examples of how to use it?[x]  **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the module 2) Report issues or problems with the module 3) Seek support
>
> ### Pedagogy / Instructional design (Work-in-progress: reviewers, please comment!)
> * [x]  **Learning objectives:** Does the module make the learning objectives plainly clear? (We don't require explicitly written learning objectives; only that they be evident from content and design.)[x]  **Content scope and length:** Is the content substantial for learning a given topic? Is the length of the module appropriate?[x]  **Pedagogy:** Does the module seem easy to follow? Does it observe guidance on cognitive load? (working memory limits of 7 +/- 2 chunks of information)[x]  **Content quality:** Is the writing of good quality, concise, engaging? Are the code components well crafted? Does the module seem complete?[x]  **Instructional design:** Is the instructional design deliberate and apparent? For example, exploit worked-example effects; effective multi-media use; low extraneous cognitive load.
>
> ### JOSE paper
> * [x]  **Authors:** Does the `paper.md` file include a list of authors with their affiliations?[x]  **A statement of need:** Does the paper clearly state the need for this module and who the target audience is?[x]  **Description:** Does the paper describe the learning materials and sequence?[x]  Does it describe how it has been used in the classroom or other settings, and how someone might adopt it?[x]  Could someone else teach with this module, given the right expertise?[x]  Does the paper tell the "story" of how the authors came to develop it, or what their expertise is?[x]  **References:** Do all archival references that should have a DOI list one (e.g., papers, datasets, software)?
