### 2025-10-27 #301 @rkn2
Okay awesome @nicoguaro  so I dont have to wait for both to come in? I can just start working on these now while we wait for the next ones? Thanks!
