### 2025-11-05 3a6c000
Review of Statics_Lesson1.ipynb Reviewer Comment (Issue/Weakness): Minor typos: "Sutdents" → Students, "analyize" → analyze. Our Response: Thank you for this catch; we have updated the text. Reviewer Comment (Issue/Weakness): Video placeholder: The code block for IPython.display. HTML does not contain a working link (currently just video_html = """ """). Replace with actual embedded video or remove. Our Response: It worked when I had downloaded the notebook but not when it was just accessible on the web interface. Since I am not sure why and could not fix this I have removed the video in favor
[…]
