### 2025-11-05 rkn2/AK_statics#1 @rkn2
@dsaninv2  for lesson 1 thanks so much for all the great suggestions! We are working through each lesson now!

**Reviewer Comment (Issue/Weakness): Minor typos: "Sutdents" → Students, "analyize" → analyze.**
*Thank you for this catch; we have updated the text.*

**Reviewer Comment (Issue/Weakness): Video placeholder: The code block for IPython.display. HTML does not contain a working link (currently just video_html = ). Replace with actual embedded video or remove.**
*It worked when I had downloaded the notebook but not when it was just accessible on the web interface. Since I am not sure why and could not fix this I have removed the video in favor of an image.*

**Reviewer Comment (Issue/Weakness): Units: Constants use PSI and inches, while students should also be familiar with SI (MPa, meters). At least provide dual-unit examples.**
*After talking with our corresponding high school teacher we decided that adding both units throughout might be a bit much for an introductory lesson. To still address this though we added a separate cell that explains how a teacher can perform the unit conversions if they want to introduce that concept.*

**Reviewer Comment (Issue/Weakness): Code readability: Some explanations are verbose. Students may benefit from concise comments plus a short "summary cell" that shows the complete workflow at once.**
*The comments were written in collaboration with the teacher who would be using the lesson. They are intended to provide comprehensive information, which teachers can then condense as needed. Therefore, we will retain the full comments but have also incorporated a summary cell, as suggested. We appreciate this valuable idea!*

**Reviewer Comment (Issue/Weakness): Learning Objectives: Reframe as measurable outcomes (e.g., "Students will be able to calculate safe ice loads using Python functions").**
*I have revised the "Learning Objectives" to reflect more measurable outcomes. This includes adding a point about students' ability to calculate safe ice loads using Python code.*

**Reviewer Comment (Issue/Weakness): Code clarity: Show a single-cell compact version after the step-by-step walkthrough, so students see the "full solution" cleanly.**
*Great idea! This is now included.*

**Reviewer Comment (Issue/Weakness): Code clarity: Use f-strings consistently for output (print(f"Ice Thickness: {h} in, Safe Load: (safe_load) lbs")).**
*Good catch! Done!*

**Reviewer Comment (Issue/Weakness): Plot improvements: Add grid lines, units on axes, and compare safe vs. theoretical load.**
*I've updated the plot to include grid lines, units on the axes, and a comparison between the safe and theoretical load capacities.*

**Reviewer Comment (Issue/Weakness): Assessment prompts: Add quick tasks at the end ("What happens if thickness = 8 inches?" or "Modify safety factor to 3, what changes?").**
*Great idea! Done!*

**Reviewer Comment (Issue/Weakness): Context connection: After calculations, explicitly ask students to reflect: "How might climate change affect the assumptions in our model?"**
*That's a great way to connect the technical calculations back to the broader context of climate change! I added a reflection prompt after the coding and plotting sections to encourage students to think about how climate change might affect the assumptions made in our model. Thanks!*
