### 2025-11-24 fe5b295
Implement reviewer feedback for Statics Lessons 1-5
