+++ Response_to_Reviewers.md
@@ -1,0 +1,510 @@
+# Response to Reviewers - AK_statics
+
+## Review of Statics_Lesson1.ipynb
+
+### Reviewer Comment (Issue/Weakness): Minor typos
+**Issue:** "Sutdents" → Students, "analyize" → analyze.
+
+**Our Response:**
+Thank you for this catch; we have updated the text.
+
+---
+
+### Reviewer Comment (Issue/Weakness): Video placeholder
+**Issue:** The code block for IPython.display.HTML does not contain a working link (currently just `video_html = """ """`). Replace with actual embedded video or remove.
+
+**Our Response:**
+It worked when I had downloaded the notebook but not when it was just accessible on the web interface. Since I am not sure why and could not fix this, I have removed the video in favor of an image.
+
+---
+
+### Reviewer Comment (Issue/Weakness): Units
+**Issue:** Constants use PSI and inches, while students should also be familiar with SI (MPa, meters). At least provide dual-unit examples.
+
+**Our Response:**
+After talking with our corresponding high school teacher, we decided that adding both units throughout might be a bit much for an introductory lesson. To still address this though, we added a separate cell that explains how a teacher can perform the unit conversions if they want to introduce that concept.
+
+---
+
+### Reviewer Comment (Issue/Weakness): Code readability
+**Issue:** Some explanations are verbose. Students may benefit from concise comments plus a short "summary cell" that shows the complete workflow at once.
+
+**Our Response:**
+The comments were written in collaboration with the teacher who would be using the lesson. They are intended to provide comprehensive information, which teachers can then condense as needed. Therefore, we will retain the full comments but have also incorporated a summary cell, as suggested. We appreciate this valuable idea!
+
+---
+
+### Reviewer Comment (Issue/Weakness): Learning Objectives
+**Issue:** Reframe as measurable outcomes (e.g., "Students will be able to calculate safe ice loads using Python functions").
+
+**Our Response:**
+I have revised the "Learning Objectives" to reflect more measurable outcomes. This includes adding a point about students' ability to calculate safe ice loads using Python code.
+
+---
+
+### Reviewer Comment (Issue/Weakness): Code clarity - compact version
+**Issue:** Show a single-cell compact version after the step-by-step walkthrough, so students see the "full solution" cleanly.
+
+**Our Response:**
+Great idea! This is now included.
+
+---
+
+### Reviewer Comment (Issue/Weakness): Code clarity - f-strings
+**Issue:** Use f-strings consistently for output (e.g., `print(f"Ice Thickness: {h} in, Safe Load: {safe_load} lbs")`).
+
+**Our Response:**
+Good catch! Done!
+
+---
+
+### Reviewer Comment (Issue/Weakness): Plot improvements
+**Issue:** Add grid lines, units on axes, and compare safe vs. theoretical load.
+
+**Our Response:**
+I've updated the plot to include grid lines, units on the axes, and a comparison between the safe and theoretical load capacities.
+
+---
+
+### Reviewer Comment (Issue/Weakness): Assessment prompts
+**Issue:** Add quick tasks at the end ("What happens if thickness = 8 inches?" or "Modify safety factor to 3, what changes?").
+
+**Our Response:**
+Great idea! Done!
+
+---
+
+### Reviewer Comment (Issue/Weakness): Context connection
+**Issue:** After calculations, explicitly ask students to reflect: "How might climate change affect the assumptions in our model?"
+
+**Our Response:**
[…]
