+++ Response_to_Reviewers.md
@@ -450,61 +450,4 @@ We've standardized the presentation of all coding tasks with clear headers indic
 
 ---
 
-## Summary of Changes
-
-### Statics_Lesson1.ipynb
-- ✅ Fixed typos
-- ✅ Replaced video with image
-- ✅ Added unit conversion cell
-- ✅ Added summary cell
-- ✅ Revised learning objectives to be measurable
-- ✅ Added compact solution version
-- ✅ Standardized f-strings
-- ✅ Enhanced plots with gridlines, units, and safe vs. theoretical comparison
-- ✅ Added assessment prompts
-- ✅ Added climate change reflection prompt
-
-### Statics_Lesson2 (1).ipynb
-- ✅ Standardized notation (FN1/FN2 throughout)
-- ✅ Fixed stability check to compare total load vs. total capacity
-- ✅ Restructured code with annotated and clean versions
-- ✅ Standardized f-strings throughout
-- ✅ Added non-symmetric example (person with backpack)
-- ✅ Added interactive coding task in concept check
-- ✅ Condensed text with worked example for system of equations
-- ✅ Added clean solution cells for each example
-
-### Statics_Lesson3.ipynb
-- ✅ Corrected all typos
-- ✅ Enhanced mathematical rigor in Example 1 (explicit reaction torque)
-- ✅ Added comments linking code to equations
-- ✅ Added intermediate print statements
-- ✅ Expanded safety discussion with explicit FoS calculation
-- ✅ Added step-by-step code outlines
-- ✅ Added FoS definition and student calculation task
-
-### Statics_Lesson4.ipynb
-- ✅ Updated vocabulary section with key definitions
-- ✅ Enhanced pivot point explanation with comparison
-- ✅ Restructured with three levels of coding tasks
-- ✅ Converted to parameterized functions
-- ✅ Added input validation and error checking
-- ✅ Improved plots with labels, gridlines, legends, and annotations
-- ✅ Clarified coding task structure
-- ✅ Added real-world failure discussion
-- ✅ Added exploration tasks section
-
-### Statics_Lesson5.ipynb
-- ✅ Added physical analogy for cut sections
-- ✅ Split into Lesson 5A (external equilibrium) and 5B (internal forces)
-- ✅ Added comprehensive vocabulary section with analogies
-- ✅ Reorganized to lead with concepts before equations
-- ✅ Provided scaffolded coding tasks (worked/partial/blank)
-- ✅ Converted to parameterized functions
-- ✅ Added interactive widgets (ipywidgets) for exploration
-- ✅ Added conceptual prediction questions
-- ✅ Standardized coding task presentation with clear headers
-
----
-
 *Document prepared: November 24, 2025*
