### 2025-11-24 fc763ed
Update Colab links for Lessons 1-5 in README
