### 2025-11-24 f7e077b
Update Colab links for Lessons 1-5 in paper.md
