### 2025-11-24 76c0491
Update Colab links to GitHub-based URLs in README and paper
