+++ paper.md
@@ -88,11 +88,11 @@ Table: Summary of course material.
 
 | Lesson | Content Summary                 | Google CoLab|
 |------|---------------------------------|------|
-| 1    | This lesson introduces the fundamental concepts of statics, focusing on the principles of forces in equilibrium. It provides an understanding of how to solve problems involving force vectors, both through graphical and analytical methods. The lesson emphasizes the importance of understanding the conditions for equilibrium in various physical scenarios, and it guides students in applying these principles to practical, real-world problems. | [![Unit_01](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/drive/1n15joLk06cVlrJ5JxutT56Ezn3lhBW6S?usp=sharing) |
-| 2    | The lesson covers concepts like support forces and net force, and teaches students how to apply the principles of vertical and horizontal equilibrium. It emphasizes visualizing parametric relationships and includes practical examples like evaluating the stability of a person standing on ice. | [![Unit_02](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/drive/1Ms4OPGwwpkaqLQ_htVPttZb3wO6UNGuU?usp=sharing) |
-| 3    | This lesson introduces the concepts of torque and rotational equilibrium. This lesson teaches students how to calculate torque and understand its role in physical systems, particularly focusing on its application in various equilibrium scenarios| [![Unit_03](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/drive/1Fz3UTplh9p3T_1X30pcbsWDhnG46tNxd?usp=sharing) |
-| 4    | This lesson reinforces understanding of forces and equilibrium in physics and engineering, with a focus on practical problem-solving.| [![Unit_04](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/drive/1wlFjsbyRVeGHSq_BDX65STMt5bZPCbVl?usp=sharing) |
-| 5    | This lesson focuses on understanding conditions for static equilibrium in bridges, including balancing forces and torques. Students learn about internal forces like shear forces and moments that contribute to equilibrium. | [![Unit_05](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/drive/1ztOiB62TMAkigSTN_lvfrreXForYGtAn?usp=sharing) |
+| 1    | This lesson introduces the fundamental concepts of statics, focusing on the principles of forces in equilibrium. It provides an understanding of how to solve problems involving force vectors, both through graphical and analytical methods. The lesson emphasizes the importance of understanding the conditions for equilibrium in various physical scenarios, and it guides students in applying these principles to practical, real-world problems. | [![Unit_01](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/rkn2/AK_statics/blob/main/Statics_Lesson1.ipynb) |
+| 2    | The lesson covers concepts like support forces and net force, and teaches students how to apply the principles of vertical and horizontal equilibrium. It emphasizes visualizing parametric relationships and includes practical examples like evaluating the stability of a person standing on ice. | [![Unit_02](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/rkn2/AK_statics/blob/main/Statics_Lesson2%20(1).ipynb) |
[…]
