### 2025-11-25 33cc73a
Standardize 'Three Levels' coding practice and 'Exploration Tasks' across all lessons
