### 2025-11-25 f55a1f3
Remove broken video from Lesson 5 and update text
