### 2025-11-25 #301 @Tejada90
> ## Review checklist for [@Tejada90](https://github.com/Tejada90)
> ### Conflict of interest
> * [x]  As the reviewer I confirm that I have read the [JOSE conflict of interest policy](https://github.com/openjournals/joss/blob/jose/COI.md) and that there are no conflicts of interest for me to review this work.
>
> ### Code of Conduct
> * [x]  I confirm that I read and will adhere to the [JOSE code of conduct](https://jose.theoj.org/about#code_of_conduct).
>
> ### General checks
> * [x]  **Repository:** Is the source for this learning module available at the https://github.com/rkn2/AK_statics?[x]  **License:** Does the repository contain a plain-text LICENSE file with the contents of a standard license? (OSI-approved for code, Creative Commons for content)[x]  **Version:** Does the release version given match the repository release?[x]  **Authorship:** Has the submitting author ([@rkn2](https://github.com/rkn2)) made visible contributions to the module? Does the full list of authors seem appropriate and complete?
>
> ### Documentation
> * [x]  **A statement of need:** Do the authors clearly state the need for this module and who the target audience is?[ ]  **Installation instructions:** Is there a clearly stated list of dependencies?[ ]  **Usage:** Does the documentation explain how someone would adopt the module, and include examples of how to use it?[ ]  **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the module 2) Report issues or problems with the module 3) Seek support
>
> ### Pedagogy / Instructional design (Work-in-progress: reviewers, please comment!)
> * [ ]  **Learning objectives:** Does the module make the learning objectives plainly clear? (We don't require explicitly written learning objectives; only that they be evident from content and design.)[x]  **Content scope and length:** Is the content substantial for learning a given topic? Is the length of the module appropriate?[x]  **Pedagogy:** Does the module seem easy to follow? Does it observe guidance on cognitive load? (working memory limits of 7 +/- 2 chunks of information)[x]  **Content quality:** Is the writing of good quality, concise, engaging? Are the code components well crafted? Does the module seem complete?[ ]  **Instructional design:** Is the instructional design deliberate and apparent? For example, exploit worked-example effects; effective multi-media use; low extraneous cognitive load.
>
> ### JOSE paper
> * [x]  **Authors:** Does the `paper.md` file include a list of authors with their affiliations?[x]  **A statement of need:** Does the paper clearly state the need for this module and who the target audience is?[x]  **Description:** Does the paper describe the learning materials and sequence?[x]  Does it describe how it has been used in the classroom or other settings, and how someone might adopt it?[x]  Could someone else teach with this module, given the right expertise?[x]  Does the paper tell the "story" of how the authors came to develop it, or what their expertise is?[x]  **References:** Do all archival references that should have a DOI list one (e.g., papers, datasets, software)?

# **JOSE Review – AK_statics (rkn2/AK_statics) by Juan C. Tejada**

https://github.com/rkn2/AK_statics/issues/2

Thank you for submitting AK_statics to JOSE. I have completed an initial review of the repository and the accompanying paper. Overall, this is a strong and valuable learning module with clear potential for use in introductory statics instruction.

My detailed checklist and comments are included below. In general, the module requires a few adjustments related to licensing, the statement of need, learning objectives, and documentation of dependencies before it can meet JOSE’s publication standards. These should be manageable revisions.

I look forward to working with you through the review process.

**Recommendation: Minor Revisions**

---

## **Conflict of Interest**

I have reviewed the JOSE conflict of interest policy and I confirm that I do not have any conflicts regarding this submission.

## **Code of Conduct**

I have read and will follow the JOSE Code of Conduct during this review.

---

# **General Checks**

### **Repository**

The repository listed in the submission is accessible and contains the materials for the learning module.

### **License**

There is a LICENSE file in plain text and it covers the code under the MIT license, which is appropriate.
However, the learning content (notebooks, text, figures) does not appear to carry a Creative Commons license, as required for JOSE’s educational materials.
A CC-BY or similar license should be added and clarified in the repository.

### **Version**

I did not see a clearly defined release (e.g., `v1.0`) corresponding to the version listed in the submission.
A tagged release would help align the submission with JOSE expectations.

### **Authorship**

The submitting author has visible contributions in the repository, and the list of authors seems reasonable.
A brief description of author contributions would be useful but is not strictly required.

---

# **Documentation**

### **Statement of Need**

The current README provides general context, but it does not clearly explain the specific need for this module, the gap it fills, or who the intended users are.
A short “Statement of Need” section would strengthen both the repository and the paper.

### **Installation / Dependencies**

Python ≥ 3.9 is mentioned, but dependencies are not clearly listed.
There is no `requirements.txt`, `environment.yml`, or similar file.
Providing explicit dependency information is important for long-term usability.

### **Usage**

The documentation shows how to open the notebooks in Google Colab and the examples in each notebook are straightforward to follow.
This part is fine.

### **Community Guidelines**

The README briefly mentions how to contribute or raise an issue, but more structured guidelines would be helpful.
A `CONTRIBUTING.md` and (optionally) a `CODE_OF_CONDUCT.md` would improve clarity for future contributors.

---

# **Pedagogy and Instructional Design**

### **Learning Objectives**

The notebooks do not present explicit learning objectives, and they are not always evident from the structure alone.
Adding a few sentences at the start of each lesson would clarify what students are expected to learn.

### **Scope and Length**

The length and scope are appropriate for an introductory treatment of statics using computational tools.

### **Pedagogical Approach**

The notebooks are generally clear and beginner-friendly.
Some lessons could benefit from slightly more scaffolding or brief explanations of why certain steps are taken.
End-of-lesson summaries or reflection questions would also help reinforce the learning process.

### **Content Quality**

The writing is accessible and the examples are relevant.
The real-world context is a strength.
Some refinement to the instructional flow would further improve the learning experience.

### **Instructional Design**

The structure suggests an intuitive progression, but some of the instructional design principles could be made more explicit—for example, including worked examples or notes to help manage cognitive load.

---

# **JOSE Paper (`paper.md`)**

### **Authors**

The paper lists its authors, it's ok.

### **Statement of Need**

As with the repository, the paper needs a clearer and more explicit statement on why this module was created, who it is intended for, and what educational gap it addresses.

### **Description of the Module**

The paper briefly describes the contents, but it does not say how the module has been used in teaching settings, nor does it guide others on how they might adopt it.
A short explanation of classroom usage, prerequisites, or typical workflows would be valuable.

### **Teachability**

A reader with the right background could use the materials, but the paper would benefit from clearer guidance for instructors who might want to adopt the module.

### **Author Background / Motivation**

It would help the reader to know a bit more about the authors' perspective or motivation for developing the module. This is not required, but it does contribute to the narrative expected in JOSE submissions.

### **References**

Any cited papers, datasets, or software packages that have DOIs should include them.

---

# **Overall Assessment**

This module has a solid foundation and clear educational value. The content is relevant, well contextualized, and accessible for learners. The computational approach using Colab is also very convenient for instructors and students.

Before the submission can be accepted, several points need attention:

1. Add a Creative Commons license for the educational content.
2. Provide a clearer Statement of Need (in the README and the paper).
3. Include explicit learning objectives in the notebooks.
4. List dependencies clearly.
5. Update paper.md with affiliations and expanded guidance for instructors.
6. Add DOIs to all references where applicable.

These changes fall under **minor revisions** and should be straightforward for the authors to address.

Once these adjustments are made, the module should be ready for acceptance into JOSE.

In summary, AK_statics presents a solid educational resource with practical examples and well-structured notebooks. The core content is sound and well aligned with computationally-enabled teaching. Before acceptance, the submission needs minor revisions in four main areas: adding a CC license for the learning materials, strengthening the statement of need, clarifying dependencies, and improving both the learning objectives and the paper’s instructor-guidance sections. Once these adjustments are addressed, the module will be in good shape for publication.
