### 2025-11-25 #301 @rkn2
thanks @Tejada90  I will work on these this week!!
