### 2025-11-25 48cfe0b
Address JOSE review: add docs, license info, dependencies, and paper updates
