+++ paper.md
@@ -46,6 +46,8 @@ This course module intertwines physics education with the unique challenges face
 This course module, designed for first-year high school students with no prior knowledge of static equilibrium and programmatic data analysis, serves as an introductory bridge to programming through high school physics. It is structured as a multi-week course that introduces Python programming and its application in physics. The module comprises five units tailored for an in-person learning environment, featuring live coding sessions and explanations, followed by independent or pair coding activities. Hosted on CoLab, it supports both self-guided and classroom-based learning, extending educational opportunities beyond the traditional classroom setting.
 
 # Statement of Need
+Engineering statics is a foundational course for civil and mechanical engineering students, yet many struggle to connect abstract mathematical concepts with physical reality. Traditional teaching methods often rely on static diagrams and manual calculations, which can obscure the dynamic relationship between forces, geometry, and equilibrium. `AK_statics` addresses this gap by providing an interactive, computational approach to learning statics. By using Python and Jupyter Notebooks, students can visualize forces in real-time and experiment with parameters to see immediate effects on stability. This module is designed for undergraduate engineering students and educators looking for open-source active learning resources.
+
 ## Contextualized STEM Content
 The benefits of contextualized science, technology, engineering, and mathematics (STEM) teaching are well known. "Contextualized Modules in Physics for Junior High School Students" by Marzan indicated that contextualized modules in physics enabled students to outperform the traditional lecture group in post-test evaluations [@marzan2018contextualized]. Similar results were seen by Cid in "Contextualized magnetism in secondary school: learning from the LHC (CERN)" [@cid2005contextualized]. That article discussed the introduction of simple physical calculations related to magnetic phenomena in old accelerators and the Large Hadron Collider. They discussed how these calculations can be used in classrooms to spark students' curiosity, aid their understanding of physical concepts, and illustrate the connection between theoretical physics ("cold equations") and the dynamic world of scientific research.
 Furthering the discussion about contextualized education, Cech's article in Nature, titled "Education: Embed social awareness in science curricula," emphasizes the need for a more integrated approach to teaching socially responsible students science, technology, engineering, and mathematics (STEM) in general [@cech2014education]. The study conducted by Cech showed that engineering students often leave university less interested in public welfare than when they started, which is contrary to what should be happening. The article suggests that STEM education should incorporate social issues throughout its curricula, not just in separate ethics courses. This is based on the premise that STEM practitioners who understand the role of their profession in society are better equipped to solve real-world problems. The article concludes that addressing public-welfare concerns should be a fundamental part of STEM education, suggesting that even a small proportion of homework and exam questions focusing on the social implications of scientific work could help reverse the trend of disengagement from these crucial issues. This approach is vital for producing well-rounded professionals capable of addressing the complex challenges of our time. 
@@ -97,6 +99,16 @@ Table: Summary of course material.
 
 ## Experience of Use
[…]
