+++ paper.bib
@@ -6,7 +6,8 @@
   number={3},
   pages={609--638},
   year={2021},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/S11191-020-00191-X}
 }
 
 @article{marzan2018contextualized,
@@ -15,7 +16,8 @@
   journal={The Vector: International Journal of Emerging Science, Technology and Management (IJESTM)},
   volume={27},
   number={1},
-  year={2018}
+  year={2018},
+  doi={10.69566/ijestm.v27i1.37}
 }
 
 @article{cid2005contextualized,
@@ -26,7 +28,8 @@
   number={4},
   pages={332},
   year={2005},
-  publisher={IOP Publishing}
+  publisher={IOP Publishing},
+  doi={10.1088/0031-9120/40/4/002}
 }
 @article{cech2014education,
   title={Education: Embed social awareness in science curricula},
@@ -36,7 +39,8 @@
   number={7484},
   pages={477--478},
   year={2014},
-  publisher={Nature Publishing Group UK London}
+  publisher={Nature Publishing Group UK London},
+  doi={10.1038/505477a}
 }
 @inproceedings{fila2014people,
   title={The people part of engineering: Engineering for, with, and as people},
@@ -44,7 +48,8 @@
   booktitle={2014 IEEE Frontiers in Education Conference (FIE) Proceedings},
   pages={1--9},
   year={2014},
-  organization={IEEE}
+  organization={IEEE},
+  doi={10.1109/FIE.2014.7044106}
 }
 
 @article{upegui2022integration,
@@ -55,7 +60,8 @@
   number={4},
   pages={923--941},
   year={2022},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s11191-021-00287-y}
 }
 
 @misc{us_census_bureau_2010,
@@ -75,7 +81,8 @@
   number={3},
   pages={633--633},
   year={1996},
-  publisher={Cambridge University Press}
+  publisher={Cambridge University Press},
+  doi={10.2307/281872}
 }
 
 @article{holen2012subsistence,
