+++ requirements.txt
@@ -1,0 +1,4 @@
+numpy>=1.20.0
+matplotlib>=3.4.0
+ipywidgets>=7.6.0
+jupyter>=1.0.0
