+++ CONTRIBUTING.md
@@ -1,0 +1,31 @@
+# Contributing to AK_statics
+
+Thank you for your interest in contributing to AK_statics! We welcome contributions from the community to help improve these educational resources.
+
+## How to Contribute
+
+### Reporting Issues
+If you find a bug, a typo, or have a suggestion for improvement, please open an issue in the GitHub repository.
+1. Check existing issues to see if it has already been reported.
+2. Open a new issue with a clear title and description.
+3. Include steps to reproduce the issue if applicable.
+
+### Suggesting Enhancements
+We welcome ideas for new lessons, examples, or features. Please open an issue to discuss your idea before submitting a pull request.
+
+### Pull Requests
+1. Fork the repository.
+2. Create a new branch for your feature or fix (`git checkout -b feature/AmazingFeature`).
+3. Make your changes.
+4. Ensure all notebooks execute without errors.
+5. Commit your changes (`git commit -m 'Add some AmazingFeature'`).
+6. Push to the branch (`git push origin feature/AmazingFeature`).
+7. Open a Pull Request.
+
+## Style Guide
+- **Notebooks**: Ensure all cells are cleared of output before committing (unless output is necessary for understanding).
+- **Code**: Follow PEP 8 guidelines for Python code.
+- **Content**: Maintain the clear, accessible tone used throughout the modules.
+
+## License
+By contributing, you agree that your contributions will be licensed under the project's licenses (MIT for code, CC-BY 4.0 for content).

[…]
