### 2025-11-25 314f5ca
Format JOSE review response to match style and remove emojis
