### 2025-12-10 #301 @nicoguaro
@rkn2, thanks for addressing the comments by both reviewers.

@dsaninv2, I can see that you have all your checkboxes marked. Would you recommend publication at this point?

@Tejada90, can you check if the changes that were made address your points and mark the boxes (if that's the case).
