### 2026-02-10 #301 @Tejada90
> [@rkn2](https://github.com/rkn2), thanks for addressing the comments by both reviewers.
>
> [@dsaninv2](https://github.com/dsaninv2), I can see that you have all your checkboxes marked. Would you recommend publication at this point?
>
> [@Tejada90](https://github.com/Tejada90), can you check if the changes that were made address your points and mark the boxes (if that's the case).

@nicoguaro I have examined the revised manuscript and confirmed that the authors have appropriately incorporated the suggested changes. These revisions effectively resolve the concerns raised in my earlier review. In light of the current version of the manuscript, I support moving forward with its publication.
