### 2026-02-11 #301 @nicoguaro
@labarba, this paper is ready to move forward.
