### 2026-02-11 #301 @labarba
@nicoguaro – Do go ahead and run `@editorial recommend-accept` if both reviewers are finished and satisfied with the revisions. Also runs `@editorialbot create post-review checklist` and ask the authors to check of their items!
