### 2026-02-11 #301 @nicoguaro
@rkn2, we are missing the archive of the repo. Can you do it and come back with the Zenodo DOI?
