### 2026-04-15 #301 @nicoguaro
@rkn2, would you mind changing the badge as suggested?

> Difficult to tell for sure, but I suspect it's these images in the tables `https://colab.research.google.com/assets/colab-badge.svg` and them not being allowed as linked items in tables in JATS XML. I would suggest trying to turn these into simple text links and see if that fixes it.
