### 2026-04-15 #301 @editorialbot
:wave: @openjournals/jose-eics, this paper is ready to be accepted and published.
