### 2026-04-16 #301 @rkn2
Yes I think its all set now!

From: Nicolás Guarín-Zapata ***@***.***>
Date: Wednesday, April 15, 2026 at 2:12 PM
To: openjournals/jose-reviews ***@***.***>
Cc: Napolitano, Rebecca ***@***.***>, Mention ***@***.***>
Subject: Re: [openjournals/jose-reviews] [REVIEW]: Static Equilibrium in Remote Alaska (Issue #301)

[https://avatars.githubusercontent.com/u/1097787?s=20&v=4]nicoguaro left a comment (openjournals/jose-reviews#301)<https://github.com/openjournals/jose-reviews/issues/301#issuecomment-4254391473>

@rkn2<https://github.com/rkn2>, could you see the checklist above?
