### 2026-05-20 #301 @labarba
I have noticed the following sentence in the paper:
> In this perspective piece, we shed light on the intricate relationship between social science and engineering challenges
faced in building and maintaining bridges in remote Alaska.

This sounds like it may be text intended for a different paper? In fact, there are various passages that don't quite belong in a JOSE paper, meant to describe the learning artifacts. I don't see this addressed in the reviewer comments, and I realize they have approved of the paper for publication, but given the sentence flagged above, perhaps you might trim a bit the excess background material...
